test: add tests for nested 'listen' blocks
These tests intend to illustrate the serial behaviour of nested 'listen'
blocks. As expected, the tests show that the nested block is dependent
on the outer block trigger to run. These tests show how nested listen
blocks work like a state machine where only 1 listen is actually
'active' at a time:

	    Start
	      |
	      V
	Outer listen <--------,
	      |               |
	      V               |
	Outer triggered --No--'
	      |
	     Yes
	      |
	      V
	Inner listen ...
	      :

Since the listen block is always evaluated once upon entry, it can still
behave as well in some asynchronous scenarios in which the inner
listen's return condition is met before the outer one. This is not true
in the general case of return conditions being met only sporadically.
To ensure such a block returns, one must ensure that all the return
conditions are met at a certain point in time and / or that they are met
in order from the outer most one to the inner most one.

This should be intuitive enough as the 'listen' block is only supposed
to run on its own trigger and therefore makes no difference that the
triggered block has itself a trigger of its own.